Development team comprehensive requirement - move LandlordFastController content to LandlordController and remove it - change route in landlords/table.blade.php - change route in routes/web.php
